Refactor front end folder structure
