+++ packages/frontend/src/styles/StatsPage.module.css
